+++ build.gradle.kts
@@ -14,6 +14,7 @@ minecraft_fp {
 
     mixin {
         pkg = "mixin.mixins"
+        debug = true
     }
 
     core {
