+++ src/main/java/com/falsepattern/endlessids/mixin/mixins/common/blockitem/common/PacketBufferMixin.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2022-2025 FalsePattern, The MEGA Team
  * All Rights Reserved
  *
- * The above copyright notice, this permission notice and the words "MEGA"
+ * The above copyright notice, this permission notice and the word "MEGA"
  * shall be included in all copies or substantial portions of the Software.
  *
  * This program is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.falsepattern.endlessids.mixin.mixins.common.blockitem.vanilla;
+package com.falsepattern.endlessids.mixin.mixins.common.blockitem.common;
 
 import io.netty.buffer.ByteBuf;
 import org.spongepowered.asm.mixin.Mixin;
