+++ src/main/java/com/falsepattern/endlessids/mixin/mixins/common/blockitem/common/StatListMixin.java
@@ -4,7 +4,7 @@
  * Copyright (C) 2022-2025 FalsePattern, The MEGA Team
  * All Rights Reserved
  *
- * The above copyright notice, this permission notice and the words "MEGA"
+ * The above copyright notice, this permission notice and the word "MEGA"
  * shall be included in all copies or substantial portions of the Software.
  *
  * This program is free software: you can redistribute it and/or modify
@@ -20,7 +20,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.falsepattern.endlessids.mixin.mixins.common.blockitem.vanilla;
+package com.falsepattern.endlessids.mixin.mixins.common.blockitem.common;
 
 import com.falsepattern.endlessids.constants.ExtendedConstants;
 import com.falsepattern.endlessids.constants.VanillaConstants;
@@ -28,15 +28,22 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
-import net.minecraft.server.management.ItemInWorldManager;
+import net.minecraft.stats.StatList;
 
-@Mixin(ItemInWorldManager.class)
-public abstract class ItemInWorldManagerMixin {
-    @ModifyConstant(method = "tryHarvestBlock",
-                    constant = @Constant(intValue = VanillaConstants.bitsPerID,
-                                         ordinal = 0),
+@Mixin(StatList.class)
+public abstract class StatListMixin {
+    @ModifyConstant(method = "<clinit>",
+                    constant = {@Constant(intValue = VanillaConstants.blockIDCount)},
                     require = 1)
-    private int extend1(int constant) {
-        return ExtendedConstants.bitsPerID;
+    private static int modifyMineBlockStatArraySize(int constant) {
+        return ExtendedConstants.blockIDCount;
     }
+
+    @ModifyConstant(method = "<clinit>",
+                    constant = {@Constant(intValue = VanillaConstants.itemIDCount)},
+                    require = 3)
+    private static int modifyItemStatArraysSize(int constant) {
+        return ExtendedConstants.itemIDCount;
+    }
+
 }
