+++ src/main/java/com/falsepattern/endlessids/mixin/plugin/Mixin.java
@@ -419,23 +419,25 @@ public enum Mixin implements IMixins {
 
     // region BlockItem
     
-    BlockItem_Thermos(Phase.EARLY,
-                      () -> { return Ext.BlockItem.getAsBoolean() && !hasThermos(); },
-                      common("blockitem.thermos.ExtendedBlockStorageMixin")),
-    
     BlockItem_Vanilla(Phase.EARLY,
+                      () -> { return Ext.BlockItem.getAsBoolean() && !hasThermos(); },
+                      common("blockitem.vanilla.ExtendedBlockStorageMixin")),
+    BlockItem_Thermos(Phase.EARLY,
+                      () -> { return Ext.BlockItem.getAsBoolean() && hasThermos(); },
+                      common("blockitem.thermos.NibbleArrayMixin")),
+    
+    BlockItem_Common(Phase.EARLY,
                       Ext.BlockItem,
-                      common("blockitem.vanilla.BlockFireMixin",
-                             "blockitem.vanilla.BlockMixin",
-                             "blockitem.vanilla.ExtendedBlockStorageMixin",
-                             "blockitem.vanilla.ItemInWorldManagerMixin",
-                             "blockitem.vanilla.ItemStackMixin",
-                             "blockitem.vanilla.PacketBufferMixin",
-                             "blockitem.vanilla.S22PacketMultiBlockChangeMixin",
-                             "blockitem.vanilla.S24PacketBlockActionMixin",
-                             "blockitem.vanilla.StatListMixin",
-                             "blockitem.vanilla.WorldMixin",
-                             "blockitem.thermos.NibbleArrayMixin"),
+                      common("blockitem.common.BlockFireMixin",
+                             "blockitem.common.BlockMixin",
+                             "blockitem.common.ExtendedBlockStorageMixin",
+                             "blockitem.common.ItemInWorldManagerMixin",
+                             "blockitem.common.ItemStackMixin",
+                             "blockitem.common.PacketBufferMixin",
+                             "blockitem.common.S22PacketMultiBlockChangeMixin",
+                             "blockitem.common.S24PacketBlockActionMixin",
+                             "blockitem.common.StatListMixin",
+                             "blockitem.common.WorldMixin"),
                       client("blockitem.vanilla.PlayerControllerMPMixin",
                              "blockitem.vanilla.RenderGlobalMixin")),
